misc fixes (#3166)
* vk: fix buffer usage

* fix compile

+++ native/cocos/renderer/gfx-gles3/GLES3CommandBuffer.cpp
@@ -369,7 +369,7 @@ void GLES3CommandBuffer::BindStates() {
     vector<uint> &dynamicOffsetOffsets = _curGPUPipelineState->gpuPipelineLayout->dynamicOffsetOffsets;
     cmd->dynamicOffsets.resize(_curGPUPipelineState->gpuPipelineLayout->dynamicOffsetCount);
     for (size_t i = 0u; i < _curDynamicOffsets.size(); i++) {
-        uint count = dynamicOffsetOffsets[i + 1] - dynamicOffsetOffsets[i];
+        size_t count = dynamicOffsetOffsets[i + 1] - dynamicOffsetOffsets[i];
         //CCASSERT(_curDynamicOffsets[i].size() >= count, "missing dynamic offsets?");
         count = std::min(count, _curDynamicOffsets[i].size());
         if (count) memcpy(&cmd->dynamicOffsets[dynamicOffsetOffsets[i]], _curDynamicOffsets[i].data(), count * sizeof(uint));
